% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_formula_gen.R
\name{bayes_formula_gen}
\alias{bayes_formula_gen}
\title{combines terms into formula}
\usage{
bayes_formula_gen(dvname, conditions = NULL, groups, mmtype = "intercept")
}
\value{
Returns bayes formulae
}
\description{
combines terms into formula
}
