% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_mm_interact.R
\name{bayes_mm_interact}
\alias{bayes_mm_interact}
\title{Output plot and means table for interactions. Can plot discrete or continuous factors. Can be passed plotting variables for pirateye}
\usage{
bayes_mm_interact(
  bmm,
  interacts = NULL,
  binom = F,
  title = NULL,
  pred_values = NULL,
  ribbon = T,
  ladder = F,
  spark = F,
  slopes = T,
  mfx_sideplot = F,
  ylim = NULL,
  ...
)
}
\arguments{
\item{bmm}{needs a bayes mixed model object (or a list with one at start)}

\item{interacts}{interactions you want to plot with : separating, gives all if void}

\item{pred_values}{for continuous variables, a vector of values where predictions are made. Defaults to 10 equally spaced steps}

\item{ribbon}{shades area where no evidence lines are different}

\item{ladder}{writes MPEs at each predicted value}

\item{spark}{plots the MPEs at each predicted value below main plot}

\item{mfx_sideplot}{puts a density plot of main effects to the side of a continuous variable}

\item{...}{pirateye plotting parameters}
}
\description{
Output plot and means table for interactions. Can plot discrete or continuous factors. Can be passed plotting variables for pirateye
}
