% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_mm_run.R
\name{bayes_mm_run}
\alias{bayes_mm_run}
\title{Runs Bayesian hierarchical mixed model with up to three conditions
and nested factors (groups).Uses rstanarm}
\usage{
bayes_mm_run(
  data,
  dvname,
  conditions,
  groups,
  f = NULL,
  mmtype = "intercept",
  binom = F,
  adapt_delta = 0.8,
  shiny = F,
  iter = 2000
)
}
\arguments{
\item{data}{data in long format}

\item{dvname}{column name for DV}

\item{conditions}{up to 3 conditions}

\item{groups}{a set of random factors}

\item{f}{formula, if omitted this will be generated for you}

\item{mmtype}{default to random intercepts, but can have 'slopes' too}

\item{binom}{If true uses the binomial / logit family}

\item{adapt_delta}{defults to 0.8, but can get closer to 1 if convergences probs}

\item{iter}{iterations for the model to run}
}
\value{
Returns the bayes model
}
\description{
Runs Bayesian hierarchical mixed model with up to three conditions
and nested factors (groups).Uses rstanarm
}
