% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_punctuationclean.R
\name{bayes_punctuationclean}
\alias{bayes_punctuationclean}
\title{cleans dt of punctuation from levels and colnames specified in cvec, replacing with _
or cvec itself if no dt}
\usage{
bayes_punctuationclean(dt = NULL, cvec, subpat = "[[:punct:]|[:space:]]")
}
\description{
cleans dt of punctuation from levels and colnames specified in cvec, replacing with _
or cvec itself if no dt
}
