% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyethinkbayes-package.R
\docType{package}
\name{eyethinkbayes-package}
\alias{eyethinkbayes}
\alias{eyethinkbayes-package}
\title{eyethinkbayes: Bayesian analysis for eyethink lab data}
\description{
Runs bayesian mixed models, bayes ttests and some bayes factors. Outputs results to text files and plots
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dcr-eyethink/eyethinkbayes}
  \item \url{https://dcr-eyethink.github.io/eyethinkbayes/}
  \item Report bugs at \url{https://github.com/dcr-eyethink/eyethinkbayes/issues}
}

}
\keyword{internal}
