% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_mm_interact.R
\name{bayes_mm_interact}
\alias{bayes_mm_interact}
\title{Output plot and means table for interactions. Can plot discrete or continuous factors. Can be passed plotting variables for pirateye}
\usage{
bayes_mm_interact(
  bmm,
  interacts = NULL,
  binom = F,
  title = NULL,
  pred_values = NULL,
  ribbon = T,
  ladder = F,
  spark = F,
  slope_mpes = T,
  mfx_sideplot = F,
  ylim = NULL,
  contrast_sparkline = F,
  show_mpe = T,
  show_preddist = T,
  ...
)
}
\arguments{
\item{bmm}{needs a bayes mixed model object (or a list with one at start)}

\item{interacts}{interactions you want to plot with : separating, gives all if void. Remember that punctuation gets stripped from condition names}

\item{pred_values}{for continuous variables, a vector of values where predictions are made. Defaults to 10 equally spaced steps}

\item{ribbon}{for continuous variables, shades area where no evidence lines are different}

\item{ladder}{for continuous variables, writes MPEs at each predicted value}

\item{spark}{for continuous variables,plots the MPEs at each predicted value below main plot}

\item{contrast_sparkline}{show distribution of contrast estimate below plot}

\item{show_mpe}{report the mpes between pairs of mean estimates for the first names condition}

\item{show_preddist}{show distributions for the mean estimates}

\item{...}{pirateye plotting parameters}
}
\description{
Output plot and means table for interactions. Can plot discrete or continuous factors. Can be passed plotting variables for pirateye
}
